# Performance optimization for large projects