--- a/src/storage/pglite-adapter.ts
+++ b/src/storage/pglite-adapter.ts
@@ -3423,9 +3423,9 @@ export class PGLiteStorageAdapter implements StorageAdapter {
 
   /**
    * Get direct access to the database connection for advanced operations
-   * Returns any to support both PGlite and Kysely query builder interfaces
+   * Returns PGlite instance for direct database access
    */
-  getDb(): any {
+  getDb(): PGlite {
     return this.db;
   }
 }