… package.json
… src/storage/bulk-insert-utils.ts
… src/utils/analysis-cache.ts
… src/utils/parallel-processor.ts
… src/workers/analysis-worker.ts